Move accounts module and update examples
