+++ itsim/node/__init__.py
@@ -13,10 +13,10 @@ from itsim import ITObject
 from itsim.network.link import Link
 from itsim.network.location import AddressInUse, InvalidAddress, Location
 from itsim.network.packet import Payload, Packet
-from itsim.node.accounts import UserAccount
 from itsim.node.file_system import File
 from itsim.node.process_management.process import Process
 from itsim.node.process_management.thread import Thread
+from itsim.node.user_management import UserAccount
 from itsim.simulator import Simulator
 from itsim.types import Address, AddressRepr, Port, PortRepr
